pyproject.toml
# NOTE: you have to use single-quoted strings in TOML for regular expressions.
# It's the equivalent of r-strings in Python.  Multiline strings are treated as
# verbose regular expressions by Black.  Use [ ] to denote a significant space
# character.

#[tool.black]
#line-length = 120
#target-version = ['py310']
#include = '\.pyi?$'

#extend-exclude = '''
#/(
#  # The following are specific to Black, you probably don't want those.
#  tests/files/
#  | profiling/
#  | scripts/generate_schema.py  # Uses match syntax
#)
#'''
# We use the unstable style for formatting Black itself. If you
# want bug-free formatting, you should keep this off. If you want
# stable formatting across releases, you should also keep `preview = true`
# (which is implied by this flag) off.
#unstable = false
[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.10
target-version = "py310"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []
# Enable the isort rules.
extend-select = ["I"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[project]
name = "asos-agent"
description = "Solving Tasks for Drug Discovery."
requires-python = ">=3.10"
version = "0.1.0"

dependencies = [
    "pydantic>=2.10.6",
    "anthropic==0.49.0",
    "openai (==2.6.0)",
    "google-genai==1.24.0",
    "colorlog>=6.9.0",
    "rich>=13.9.4",
    "gitingest>=0.1.3",
    "pymupdf4llm>=0.0.17",
    "aiohttp>=3.11.13",
    "trafilatura>=2.0.0",
    "mcp (>=1.9.3,<2.0.0)",
    "Deprecated>=1.2",
    "tiktoken>=0.8.0",
    "PyYAML>=6.0",
    "datasets>=3.2.0"
]

[tool.isort]
atomic = true
profile = "black"
line_length = 120
skip_gitignore = true
known_first_party = ["asos-agent", "agents"]

[tool.poetry]
name = "asos-agent"
version = "0.1.0"
description = "Solving Tasks for Drug Discovery"
authors = ["Balto Devs"]
packages = [{include = "agents", from = "src"}]


[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pydantic = "^2.10.6"
anthropic = "0.49.0"
openai = "2.6.0"
google-genai = "1.24.0"
colorlog = "^6.9.0"
rich = "^13.9.4"
gitingest = "^0.1.3"
pymupdf4llm = "^0.0.17"
aiohttp = "^3.11.13"
trafilatura = "^2.0.0"
mcp = ">=1.9.3,<2.0.0"
Deprecated = ">=1.2"
tiktoken = "^0.8.0"
PyYAML = "^6.0"
datasets = "^3.2.0"

[tool.poetry.dev-dependencies]
black = "^23.3"
ruff = "==0.6.2"
pytest = "==8.3.2"
pytest-cov = "==5.0.0"
